Change enum type to enum class type
To avoid compiler error C26812.

+++ tsMuxer/vvcStreamReader.cpp
@@ -124,7 +124,7 @@ int VVCStreamReader::getTSDescriptor(uint8_t* dstBuff, bool blurayMode, bool hdm
 
         int video_format, frame_rate_index, aspect_ratio_index;
         M2TSStreamInfo::blurayStreamParams(getFPS(), getInterlaced(), getStreamWidth(), getStreamHeight(),
-                                           (float)getStreamAR(), &video_format, &frame_rate_index, &aspect_ratio_index);
+                                           (int)getStreamAR(), &video_format, &frame_rate_index, &aspect_ratio_index);
 
         *dstBuff++ = (video_format << 4) + frame_rate_index;
         *dstBuff++ = (aspect_ratio_index << 4) + 0xf;
